+++ src/Immutable/Tuple.php
@@ -3,6 +3,7 @@
 namespace MF\Collection\Immutable;
 
 use Assert\Assertion;
+use MF\Collection\Helper\Strings;
 
 class Tuple implements ITuple
 {
@@ -45,10 +46,11 @@ class Tuple implements ITuple
             return explode(',', trim($tuple, '()'));
         })
             ->reduce(function (array $matches, string $match) use (&$cache) {
-                $isStart = in_array(mb_substr(ltrim($match), 0, 1), ['"', "'"], true)
+                $trimmedMatch = ltrim($match);
+                $isStart = (Strings::startsWith($trimmedMatch, '"') || Strings::startsWith($trimmedMatch, "'"))
                     && empty($cache);
-                $isEnd = in_array(mb_substr($match, -1, 1), ['"', "'"], true)
-                    && (mb_strlen(ltrim($match)) > 1 || !empty($cache));
+                $isEnd = (Strings::endsWith($match, '"') || Strings::endsWith($match, "'"))
+                    && (mb_strlen($trimmedMatch) > 1 || !empty($cache));
 
                 if (!$isEnd && ($isStart || !empty($cache))) {
                     $cache[] = $match;
@@ -71,22 +73,7 @@ class Tuple implements ITuple
                 return trim($match);
             })
             ->filter('($match) => $match !== ""')
-            ->map(function (string $item) {
-                $item = trim($item);
-                if (is_numeric($item) && mb_strpos($item, '.') === false) {
-                    return (int) $item;
-                } elseif (is_numeric($item)) {
-                    return (float) $item;
-                } elseif ($item === 'true') {
-                    return true;
-                } elseif ($item === 'false') {
-                    return false;
-                } elseif ($item === 'null') {
-                    return null;
-                }
-
-                return preg_replace('/^[\'\"]?/', '', preg_replace('/[\'\"]?$/', '', $item));
-            })
+            ->map(self::mapParsedItem())
             ->toArray();
 
         if ($expectedItemsCount !== null) {
@@ -98,6 +85,34 @@ class Tuple implements ITuple
         }
 
         return new self($values);
+    }
+
+    private static function mapParsedItem(): callable
+    {
+        return function (string $item) {
+            $item = trim($item);
+            if (is_numeric($item) && mb_strpos($item, '.') === false) {
+                return (int) $item;
+            } elseif (is_numeric($item)) {
+                return (float) $item;
+            } elseif ($item === 'true') {
+                return true;
+            } elseif ($item === 'false') {
+                return false;
+            } elseif ($item === 'null') {
+                return null;
+            } elseif (Strings::startsWith($item, '[') && Strings::endsWith($item, ']')) {
+                $arrayContent = trim($item, '[]');
+                Assertion::false(
+                    Strings::contains($arrayContent, '[') || Strings::contains($arrayContent, ']'),
+                    sprintf('Tuple must NOT contain multi-dimensional arrays. Invalid item: "%s"', $item)
+                );
+
+                return array_map(self::mapParsedItem(), explode(';', $arrayContent));
+            }
+
+            return preg_replace('/^[\'\"]?/', '', preg_replace('/[\'\"]?$/', '', $item));
+        };
     }
 
     /**
@@ -246,19 +261,26 @@ class Tuple implements ITuple
      */
     public function toString(): string
     {
-        return sprintf('(%s)', implode(', ', array_map(function ($value) {
+        return sprintf('(%s)', implode(', ', array_map($this->mapValueToString(), $this->values)));
+    }
+
+    private function mapValueToString(): callable
+    {
+        return function ($value) {
             if ($value === null) {
                 return 'null';
             } elseif ($value === true) {
                 return 'true';
             } elseif ($value === false) {
                 return 'false';
+            } elseif (is_array($value)) {
+                return sprintf('[%s]', implode('; ', array_map($this->mapValueToString(), $value)));
             }
 
             return is_string($value)
                 ? sprintf('"%s"', $value)
                 : $value;
-        }, $this->values)));
+        };
     }
 
     public function count(): int
@@ -413,7 +435,7 @@ class Tuple implements ITuple
     {
         return function (string $type) {
             $types = Seq::create(explode('|', $type), function (string $type): iterable {
-                if (mb_substr($type, 0, 1) === '?') {
+                if (Strings::startsWith($type, '?')) {
                     yield 'NULL';
                 }