+++ tests/Immutable/MapTest.php
@@ -25,7 +25,6 @@ class MapTest extends AbstractTestCase
     }
 
     /**
-     * @param array $array
      * @param bool $recursive
      *
      * @dataProvider arrayProvider
@@ -207,8 +206,6 @@ class MapTest extends AbstractTestCase
     }
 
     /**
-     * @param array $array
-     *
      * @dataProvider arrayProvider
      */
     public function testShouldGetCount(array $array): void
@@ -386,8 +383,6 @@ class MapTest extends AbstractTestCase
     }
 
     /**
-     * @param callable $reducer
-     * @param array $values
      * @param mixed $expected
      *
      * @dataProvider reduceProvider
@@ -424,8 +419,6 @@ class MapTest extends AbstractTestCase
     }
 
     /**
-     * @param callable $reducer
-     * @param array $values
      * @param mixed $initialValue
      * @param mixed $expected
      *