+++ CHANGELOG.md
@@ -2,9 +2,11 @@
 
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
+- Supports `array` in `Tuple`
 
 ## 3.11.0 - 2018-08-17
 - Add `merge` and `mergeMatch` methods to `ITuple`
+- Remove superfluous type annotations
 
 ## 3.10.0 - 2018-07-26
 - Change return type of `getIterator` method to generic `iterable`